requirements.txt
# ═══════════════════════════════════════════════════════════
# 工业设备管理系统 - Python依赖
# Python 3.10+
# ═══════════════════════════════════════════════════════════

# ── 核心框架 ──
PySide6>=6.6.0
PySide6-Addons>=6.6.0

# ── 数据可视化 ──
pyqtgraph>=0.13.0

# ── 数据导出 ──
openpyxl>=3.1.0
csvkit>=2.4.0

# ── 串口通信 ──
pyserial>=3.5

# ── 配置管理 ──
toml>=0.10.2

# ── 开发工具（仅开发环境）──
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
